+++ src/dicognito/element_anonymizer.py
@@ -1,7 +1,11 @@
+from typing import Iterator
 from pydicom import DataElement
 from pydicom.dataset import Dataset
 
 
 class ElementAnonymizer:
     def __call__(self, dataset: Dataset, data_element: DataElement) -> bool:
+        raise NotImplementedError
+
+    def describe_actions(self) -> Iterator[str]:
         raise NotImplementedError