+++ src/dicognito/pnanonymizer.py
@@ -1,6 +1,6 @@
 import pydicom
 
-from typing import Any, Optional
+from typing import Any, Iterator, Optional
 from dicognito.element_anonymizer import ElementAnonymizer
 from dicognito.randomizer import Randomizer
 
@@ -45,6 +45,9 @@ class PNAnonymizer(ElementAnonymizer):
         else:
             data_element.value = self._new_pn(patient_sex, data_element.value)
         return True
+
+    def describe_actions(self) -> Iterator[str]:
+        yield "Replace all PN attributes with anonymized values"
 
     def _new_pn(self, sex: Optional[str], original_value: Any) -> str:
         if sex == "F":